backend/requirements.txt
# FastAPI and Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
google-cloud-bigquery==3.13.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.25.2
polars==0.20.2

# API
graphene==3.3
graphene-sqlalchemy==3.0.0
requests==2.31.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
celery==5.3.4
redis==5.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
